Create subpackage for annotators.
