+++ protokt-codegen/src/main/kotlin/com/toasttab/protokt/codegen/impl/Implements.kt
@@ -16,7 +16,7 @@
 package com.toasttab.protokt.codegen.impl
 
 import arrow.core.memoize
-import com.toasttab.protokt.codegen.impl.Annotator.Context
+import com.toasttab.protokt.codegen.annotators.Annotator.Context
 import com.toasttab.protokt.codegen.impl.ClassLookup.getClass
 import com.toasttab.protokt.codegen.model.PClass
 import com.toasttab.protokt.codegen.model.possiblyQualify
